# openapi.yaml
openapi: 3.1.0
info:
  title: RescueBite API
  description: Production-grade RescueBite backend with restaurant operations, inventory
    expiry, reservations, and manual seller-confirmed payments.
  version: 1.1.0
paths:
...
  /donations/claims:
    get:
      tags:
      - donations
      summary: List Claims
      operationId: list_claims_donations_claims_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DonationClaimRead'
                type: array
                title: Response List Claims Donations Claims Get
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Standard 422 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - donations
      summary: Create Claim
      operationId: create_claim_donations_claims_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonationClaimCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationClaimRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /donations/claims/{claimId}:
    patch:
      tags:
      - donations
      summary: Update Claim
      operationId: update_claim_donations_claims__claimId__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: integer
          title: Claimid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonationClaimUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationClaimRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - donations
      summary: Delete Claim
      operationId: delete_claim_donations_claims__claimId__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: integer
          title: Claimid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /donations/claims/{claimId}/receive:
    post:
      tags:
      - donations
      summary: Receive Claim
      operationId: receive_claim_donations_claims__claimId__receive_post
      security:
      - HTTPBearer: []
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: integer
          title: Claimid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationClaimRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ClaimStatus:
      type: string
      enum:
      - pending
      - approved
      - received
      - cancelled
      - deleted
      title: ClaimStatus
    DonationClaimCreate:
      properties:
        batch_id:
          type: integer
          title: Batch Id
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
      type: object
      required:
      - batch_id
      - quantity
      title: DonationClaimCreate
    DonationClaimRead:
      properties:
        id:
          type: integer
          title: Id
        shelter_id:
          type: integer
          title: Shelter Id
        batch_id:
          type: integer
          title: Batch Id
        quantity:
          type: integer
          title: Quantity
        status:
          $ref: '#/components/schemas/ClaimStatus'
      type: object
      required:
      - id
      - shelter_id
      - batch_id
      - quantity
      - status
      title: DonationClaimRead
    DonationClaimUpdate:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/ClaimStatus'
          - type: 'null'
      type: object
      title: DonationClaimUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      example:
        code: forbidden
        message: insufficient role
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT